SpeechComponentクラスを追加した #100
